+++ Changelog.md
@@ -2,6 +2,8 @@ v0.0.4
 ======
 - spinlocks in memory cache.
 - B+tree based caches
+- flush data on write to page
+
 v0.0.3
 ======
 - speed up